+++ src/Server/src/Core/Configuration/IdOpsServerBuilderExtensions.cs
@@ -72,14 +72,6 @@ namespace IdOps
             services.AddSingleton<IResourceStores>(p => new ResourceStores(p));
             services.AddSingleton<IResourceAuthoring>(p => new ResourceAuthoring(p));
             services.AddSingleton<IResourceManager, ResourceManager>();
-            //services.AddSingleton<IResourceManager<Application>, ResourceManager<Application>>();
-            //services.AddSingleton<IResourceManager<Client>, ResourceManager<Client>>();
-            //services.AddSingleton<IResourceManager<IdentityResource>, ResourceManager<IdentityResource>>();
-            //services.AddSingleton<IResourceManager<Model.IdentityServer>, ResourceManager<Model.IdentityServer>>();
-            //services.AddSingleton<IResourceManager<PersonalAccessToken>, ResourceManager<PersonalAccessToken>>();
-            //services.AddSingleton<IResourceManager<UserClaimRule>, ResourceManager<UserClaimRule>>();
-            //services.AddSingleton<IResourceManager<ApiResource>, ResourceManager<ApiResource>>();
-            //services.AddSingleton<IResourceManager<ApiScope>, ResourceManager<ApiScope>>();
         }
 
         private static void AddPublishing(this IServiceCollection services)
